Move code from hpp to cpp
