+++ src/lib/CMakeLists.txt
@@ -13,6 +13,7 @@ set(
     constant_mappings.cpp
     constant_mappings.hpp
     import_export/binary.hpp
+    import_export/csv_converter.cpp
     import_export/csv_converter.hpp
     import_export/csv.hpp
     import_export/csv_parser.cpp
@@ -34,7 +35,9 @@ set(
     operators/abstract_join_operator.hpp
     operators/abstract_operator.cpp
     operators/abstract_operator.hpp
+    operators/abstract_read_only_operator.cpp
     operators/abstract_read_only_operator.hpp
+    operators/abstract_read_write_operator.cpp
     operators/abstract_read_write_operator.hpp
     operators/aggregate.cpp
     operators/aggregate.hpp
@@ -128,6 +131,7 @@ set(
     optimizer/abstract_syntax_tree/stored_table_node.hpp
     optimizer/abstract_syntax_tree/update_node.cpp
     optimizer/abstract_syntax_tree/update_node.hpp
+    optimizer/base_column_statistics.cpp
     optimizer/base_column_statistics.hpp
     optimizer/column_statistics.cpp
     optimizer/column_statistics.hpp
@@ -197,9 +201,11 @@ set(
     storage/index/adaptive_radix_tree/adaptive_radix_tree_index.hpp
     storage/index/adaptive_radix_tree/adaptive_radix_tree_nodes.cpp
     storage/index/adaptive_radix_tree/adaptive_radix_tree_nodes.hpp
+    storage/index/base_index.cpp
     storage/index/base_index.hpp
     storage/index/group_key/composite_group_key_index.cpp
     storage/index/group_key/composite_group_key_index.hpp
+    storage/index/group_key/group_key_index.cpp
     storage/index/group_key/group_key_index.hpp
     storage/index/group_key/variable_length_key_base.cpp
     storage/index/group_key/variable_length_key_base.hpp
