+++ src/lib/operators/delete.hpp
@@ -26,11 +26,6 @@ class Delete : public AbstractReadWriteOperator {
   const std::string name() const override;
   uint8_t num_in_tables() const override;
 
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
-
   void finish_commit() override;
 
  protected:
