+++ src/lib/operators/export_binary.hpp
@@ -42,11 +42,6 @@ class ExportBinary : public AbstractReadOnlyOperator {
    */
   uint8_t num_out_tables() const final;
 
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
-
  private:
   // Path of the binary file
   const std::string _filename;
