+++ src/lib/operators/import_binary.hpp
@@ -51,11 +51,6 @@ class ImportBinary : public AbstractReadOnlyOperator {
   // This operator has one table as output.
   uint8_t num_out_tables() const final;
 
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
-
  private:
   /*
    * Reads the header from the given file.
