+++ src/lib/operators/import_csv.hpp
@@ -48,11 +48,6 @@ class ImportCsv : public AbstractReadOnlyOperator {
   // This operator has one table as output.
   uint8_t num_out_tables() const override;
 
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
-
  protected:
   // Returns the table that was created from the csv file.
   std::shared_ptr<const Table> _on_execute() override;
