+++ src/lib/operators/join_hash.hpp
@@ -31,10 +31,6 @@ class JoinHash : public AbstractJoinOperator {
   const std::string name() const override;
   uint8_t num_in_tables() const override;
   uint8_t num_out_tables() const override;
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
 
  protected:
   std::shared_ptr<const Table> _on_execute() override;
