+++ src/lib/operators/rollback_records.hpp
@@ -15,10 +15,6 @@ class RollbackRecords : public AbstractReadWriteOperator {
   const std::string name() const override;
   uint8_t num_in_tables() const override;
   uint8_t num_out_tables() const override;
-  std::shared_ptr<AbstractOperator> recreate(const std::vector<AllParameterVariant>& args) const override {
-    Fail("Operator " + name() + " does not implement recreation.");
-    return {};
-  }
 
   void commit_records(const CommitID cid) override;
   void rollback_records() override;
