+++ src/lib/scheduler/topology.hpp
@@ -34,14 +34,14 @@ class Topology final {
                                                              uint32_t workers_per_node = 1);
   static std::shared_ptr<Topology> create_numa_topology(uint32_t max_num_cores = 0);
 
-  Topology(std::vector<TopologyNode>&& nodes, size_t numCpus) : _nodes(std::move(nodes)), _numCpus(numCpus) {}
+  Topology(std::vector<TopologyNode>&& nodes, size_t num_cpus) : _nodes(std::move(nodes)), _num_cpus(num_cpus) {}
 
-  const std::vector<TopologyNode>& nodes() { return _nodes; }
+  const std::vector<TopologyNode>& nodes();
 
-  size_t numCpus() const { return _numCpus; }
+  size_t num_cpus() const;
 
  private:
   std::vector<TopologyNode> _nodes;
-  size_t _numCpus;
+  size_t _num_cpus;
 };
 }  // namespace opossum
